Reorganize components to reside in its page's folders
- Refactor home page code to remove unused event listeners
