+++ src/app/home/components/account-modal/account-modal.component.html
@@ -1,0 +1,3 @@
+<p>
+  account-modal works!
+</p>
