+++ src/app/home/components/account-modal/account-modal.component.ts
@@ -1,0 +1,14 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-account-modal',
+  templateUrl: './account-modal.component.html',
+  styleUrls: ['./account-modal.component.scss'],
+})
+export class AccountModalComponent  implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {}
+
+}
