+++ src/app/home/home.module.ts
@@ -7,11 +7,11 @@ import { HomePage } from './home.page';
 import { HomePageRoutingModule } from './home-routing.module';
 import { AccountFilterPipe } from '../pipes/account-filter.pipe';
 import { NgxScannerQrcodeModule } from 'ngx-scanner-qrcode';
-import { AccountListComponent } from '../components/account-list/account-list.component';
-import { AccountDetailComponent } from '../components/account-detail/account-detail.component';
-import { CountdownTimerComponent } from '../components/countdown-timer/countdown-timer.component';
+import { AccountListComponent } from './components/account-list/account-list.component';
+import { AccountDetailComponent } from './components/account-detail/account-detail.component';
+import { CountdownTimerComponent } from './components/countdown-timer/countdown-timer.component';
 import { TranslateModule } from '@ngx-translate/core'
-import { AccountSelectModalComponent } from '../components/account-select-modal/account-select-modal.component';
+import { AccountSelectModalComponent } from './components/account-select-modal/account-select-modal.component';
 
 @NgModule({
   imports: [
